+++ tests/_support/Helper/CraftLiltPluginHelper.php
@@ -14,6 +14,7 @@ namespace Helper;
 use Codeception\Module;
 use Craft;
 use craft\queue\BaseJob;
+use craft\queue\QueueInterface;
 use lilthq\craftliltplugin\Craftliltplugin;
 use lilthq\craftliltplugin\elements\Job;
 use lilthq\craftliltplugin\records\I18NRecord;
@@ -125,7 +126,7 @@ class CraftLiltPluginHelper extends Module
 
         $key = get_class($expectedJob) . '_' . json_encode($expectedJob);
 
-        $this->assertArrayHasKey($key, $jobInfos);
+        $this->assertArrayHasKey($key, $jobInfos, 'No job found in array: ' . json_encode($jobInfos));
         $this->assertEquals($expectedJob, $jobInfos[$key]);
     }
 
@@ -135,7 +136,7 @@ class CraftLiltPluginHelper extends Module
 
         $key = get_class($expectedJob) . '_' . json_encode($expectedJob);
 
-        $this->assertArrayNotHasKey($key, $jobInfos, $message);
+        $this->assertArrayNotHasKey($key, $jobInfos, $message . ': ' . json_encode($jobInfos));
     }
 
     public function assertJobIdNotInQueue(int $jobId, string $message = ''): void
@@ -190,6 +191,31 @@ class CraftLiltPluginHelper extends Module
         }
 
         $settingRecord->value = $value;
+        $settingRecord->save();
+    }
+    public function enableOption(string $name): void
+    {
+        $this->setOption($name, 1);
+    }
+
+    public function disableOption(string $name): void
+    {
+        $this->setOption($name, 0);
+    }
+
+
+    private function setOption(string $name, int $value): void
+    {
+        $settingRecord = SettingRecord::findOne(
+            ['name' => SettingsRepository::QUEUE_DISABLE_AUTOMATIC_SYNC]
+        );
+        if (!$settingRecord) {
+            $settingRecord = new SettingRecord(
+                ['name' => SettingsRepository::QUEUE_DISABLE_AUTOMATIC_SYNC]
+            );
+        }
+
+        $settingRecord->value = (string) $value;
         $settingRecord->save();
     }
 
@@ -370,6 +396,13 @@ class CraftLiltPluginHelper extends Module
         Craft::$app->getQueue()->push($job);
 
         Craft::$app->getQueue()->run();
+    }
+
+    public function clearQueue(): void
+    {
+        /** @var QueueInterface $queue */
+        $queue = Craft::$app->getQueue();
+        $queue->releaseAll();
     }
 
     public function executeQueue(string $queueItem, array $params = []): void