+++ tests/integration/modules/FetchJobStatusFromConnectorCest.php
@@ -17,7 +17,9 @@ use lilthq\craftliltplugin\modules\FetchInstantJobTranslationsFromConnector;
 use lilthq\craftliltplugin\modules\FetchJobStatusFromConnector;
 use lilthq\craftliltplugin\modules\FetchTranslationFromConnector;
 use lilthq\craftliltplugin\modules\FetchVerifiedJobTranslationsFromConnector;
+use lilthq\craftliltplugin\parameters\CraftliltpluginParameters;
 use lilthq\craftliltplugin\records\TranslationRecord;
+use lilthq\craftliltplugin\services\repositories\SettingsRepository;
 use lilthq\craftliltplugintests\integration\AbstractIntegrationCest;
 use lilthq\tests\fixtures\EntriesFixture;
 use PHPUnit\Framework\Assert;
@@ -40,6 +42,9 @@ class FetchJobStatusFromConnectorCest extends AbstractIntegrationCest
      */
     public function testExecuteSuccessVerified(IntegrationTester $I): void
     {
+        $I->clearQueue();
+        $I->disableOption(SettingsRepository::QUEUE_DISABLE_AUTOMATIC_SYNC);
+
         Db::truncateTable(Craft::$app->queue->tableName);
 
         $user = Craft::$app->getUsers()->getUserById(1);
@@ -171,6 +176,9 @@ class FetchJobStatusFromConnectorCest extends AbstractIntegrationCest
      */
     public function testExecuteSuccessInstant(IntegrationTester $I): void
     {
+        $I->clearQueue();
+        $I->disableOption(SettingsRepository::QUEUE_DISABLE_AUTOMATIC_SYNC);
+
         Db::truncateTable(Craft::$app->queue->tableName);
 
         $user = Craft::$app->getUsers()->getUserById(1);
@@ -302,6 +310,9 @@ class FetchJobStatusFromConnectorCest extends AbstractIntegrationCest
      */
     public function testExecuteJobNotFound(IntegrationTester $I): void
     {
+        $I->clearQueue();
+        $I->disableOption(SettingsRepository::QUEUE_DISABLE_AUTOMATIC_SYNC);
+
         Db::truncateTable(Craft::$app->queue->tableName);
 
         $user = Craft::$app->getUsers()->getUserById(1);
@@ -326,6 +337,9 @@ class FetchJobStatusFromConnectorCest extends AbstractIntegrationCest
      */
     public function testExecuteSuccessProcessing(IntegrationTester $I): void
     {
+        $I->clearQueue();
+        $I->disableOption(SettingsRepository::QUEUE_DISABLE_AUTOMATIC_SYNC);
+
         Db::truncateTable(Craft::$app->queue->tableName);
 
         $user = Craft::$app->getUsers()->getUserById(1);
@@ -373,8 +387,56 @@ class FetchJobStatusFromConnectorCest extends AbstractIntegrationCest
      * @throws Exception
      * @throws ModuleException
      */
+    public function testExecuteSuccessProcessingAutomaticSyncDisabled(IntegrationTester $I): void
+    {
+        $I->clearQueue();
+        $I->enableOption(SettingsRepository::QUEUE_DISABLE_AUTOMATIC_SYNC);
+
+        Db::truncateTable(Craft::$app->queue->tableName);
+
+        $user = Craft::$app->getUsers()->getUserById(1);
+        $I->amLoggedInAs($user);
+
+        $job = $I->createJob([
+            'title' => 'Awesome test job',
+            'elementIds' => [999],
+            'targetSiteIds' => '*',
+            'sourceSiteId' => Craftliltplugin::getInstance()->languageMapper->getSiteIdByLanguage('en-US'),
+            'translationWorkflow' => SettingsResponse::LILT_TRANSLATION_WORKFLOW_INSTANT,
+            'versions' => [],
+            'authorId' => 1,
+            'liltJobId' => 777,
+        ]);
+
+        $I->expectJobGetRequest(
+            777,
+            200,
+            [
+                'status' => JobResponse::STATUS_PROCESSING
+            ]
+        );
+
+        $I->runQueue(
+            FetchJobStatusFromConnector::class,
+            [
+                'liltJobId' => 777,
+                'jobId' => $job->id,
+            ]
+        );
+
+        $totalJobs = Craft::$app->queue->getJobInfo();
+        Assert::assertCount(0, $totalJobs);
+    }
+
+    /**
+     * @throws Exception
+     * @throws ModuleException
+     */
     public function testExecuteSuccessQueued(IntegrationTester $I): void
     {
+        $I->clearQueue();
+        $I->disableOption(SettingsRepository::QUEUE_DISABLE_AUTOMATIC_SYNC);
+
         Db::truncateTable(Craft::$app->queue->tableName);
 
         $user = Craft::$app->getUsers()->getUserById(1);